refactor(contracts): centralize listening-history + IPC error-code types
The listening-history wire types were duplicated verbatim between the preload
surface (preload/types.ts) and the renderer contract (electron.d.ts), and the
IPC error-code constant maps lived in the desktop-only errors module where the
renderer could not reference the same literals.

Move both into @shiranami/contracts (ipc/history.ts, ipc/error-codes.ts) as the
single source of truth. The preload types.ts re-exports the history types; the
desktop errors.ts re-exports the code maps so existing main-process import sites
keep working unchanged.
