+++ packages/contracts/src/index.ts
@@ -12,6 +12,8 @@ export * from './ipc/debug';
 export * from './ipc/metadata';
 export * from './ipc/loudness';
 export * from './ipc/database';
+export * from './ipc/error-codes';
+export * from './ipc/history';
 export * from './ipc/system';
 export * from './ipc/storage';
 export * from './share/dto';
